Projects/Daily Market Brief/requirements.txt
anthropic>=0.51.0
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2024.1
